synthesize OOB log records from OOB reports
* When logging the reception of a message, if we can parse it
  as an rfc3464 oob report, then synthesize OOB records for
  each recipient
* Add config to allow relaying for domains for which bounces
  are received, so that we can accept their message content.
* If the domain config allows bounces but not relaying, then
  the message will not be spooled or queued after reception;
  it will be dropped with no additional logging.

+++ crates/message/src/rfc3464.rs
@@ -31,7 +31,7 @@ impl ReportAction {
 
 #[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq)]
 pub struct ReportStatus {
-    pub class: u16,
+    pub class: u8,
     pub subject: u16,
     pub detail: u16,
     pub comment: Option<String>,
